Refactor: file directory로 변경
